[#35] CommentMockData으로 분리 및 코드 정리
